+++ modules/react/common/lib/styles/errorRing.ts
@@ -53,10 +53,10 @@ export function errorRing(error?: ErrorType, theme?: EmotionCanvasTheme): CSSObj
     borderColor: errorColors.outer,
     transition: '100ms box-shadow',
     boxShadow: errorBoxShadow,
-    '&:hover, &:disabled': {
+    '&:hover, &:disabled, &.hover, &.disabled': {
       borderColor: errorColors.outer,
     },
-    '&:focus:not([disabled])': {
+    '&:focus-visible:not([disabled]), &.focus:not([disabled])': {
       borderColor: errorColors.outer,
       boxShadow: `${errorBoxShadow},
         0 0 0 2px ${colors.frenchVanilla100},