+++ modules/react/select/stories/examples/LabelPosition.tsx
@@ -24,9 +24,7 @@ export const LabelPosition = () => {
           <Select.Input />
           <Select.Popper>
             <Select.Card>
-              {model.state.items.length > 0 && (
-                <Select.List>{item => <Select.Item>{item}</Select.Item>}</Select.List>
-              )}
+              <Select.List>{item => <Select.Item>{item}</Select.Item>}</Select.List>
             </Select.Card>
           </Select.Popper>
         </FormField>